Refactor message publishing logic into higher abstraction level
