+++ analytics-module/build.gradle.kts
@@ -11,6 +11,4 @@ dependencies {
 
   implementation("com.google.cloud:google-cloud-pubsub:$googleCloudVersion")
   implementation("com.google.code.gson:gson:2.8.5")
-
-  testImplementation("com.google.api:gax-grpc:1.48.1")
 }
