+++ analytics-module/src/main/kotlin/org/ostelco/prime/analytics/publishers/DataConsumptionInfoPublisher.kt
@@ -1,13 +1,9 @@
 package org.ostelco.prime.analytics.publishers
 
-import com.google.api.core.ApiFutureCallback
-import com.google.api.core.ApiFutures
-import com.google.api.gax.rpc.ApiException
 import com.google.gson.Gson
 import com.google.protobuf.ByteString
 import com.google.pubsub.v1.PubsubMessage
 import org.ostelco.prime.analytics.ConfigRegistry
-import org.ostelco.prime.getLogger
 import org.ostelco.prime.model.DataTrafficInfo
 import java.time.Instant
 
@@ -17,7 +13,6 @@ import java.time.Instant
 object DataConsumptionInfoPublisher :
         PubSubPublisher by DelegatePubSubPublisher(topicId = ConfigRegistry.config.dataTrafficTopicId) {
 
-    private val logger by getLogger()
     private val gson = Gson()
 
     /**
@@ -49,25 +44,7 @@ object DataConsumptionInfoPublisher :
                 .build()
 
         // schedule a message to be published, messages are automatically batched
-        val future = publishPubSubMessage(pubsubMessage)
-
-        // add an asynchronous callback to handle success / failure
-        ApiFutures.addCallback(future, object : ApiFutureCallback<String> {
-
-            override fun onFailure(throwable: Throwable) {
-                if (throwable is ApiException) {
-                    // details on the API exception
-                    logger.warn("Status code: {}", throwable.statusCode.code)
-                    logger.warn("Retrying: {}", throwable.isRetryable)
-                }
-                logger.warn("Error publishing message for subscriptionAnalyticsId: {}", subscriptionAnalyticsId)
-            }
-
-            override fun onSuccess(messageId: String) {
-                // Once published, returns server-assigned message ids (unique within the topic)
-                logger.debug("Published message $messageId")
-            }
-        }, singleThreadScheduledExecutor)
+        publishPubSubMessage(pubsubMessage)
     }
 
     private fun toJson(dataTrafficInfo: DataTrafficInfo): ByteString {
