+++ analytics-module/src/main/kotlin/org/ostelco/prime/analytics/publishers/PubSubPublisher.kt
@@ -1,11 +1,10 @@
 package org.ostelco.prime.analytics.publishers
 
-import com.google.api.core.ApiFuture
 import com.google.pubsub.v1.PubsubMessage
 import io.dropwizard.lifecycle.Managed
 import java.util.concurrent.ScheduledExecutorService
 
 interface PubSubPublisher : Managed {
     var singleThreadScheduledExecutor: ScheduledExecutorService
-    fun publishPubSubMessage(pubsubMessage: PubsubMessage): ApiFuture<String>
+    fun publishPubSubMessage(pubsubMessage: PubsubMessage)
 }
